# swaggerui/html/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is the API Server for the Radix job scheduler server.",
    "title": "Radix job scheduler server.",
    "contact": {
      "url": "https://equinor.slack.com/messages/CBKM6N2JY"
    },
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/batches": {
      "post": {
        "tags": [
          "Batch"
        ],
        "summary": "Create batch",
        "operationId": "createBatch",
        "parameters": [
          {
            "description": "Batch to create",
            "name": "batchCreation",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchScheduleDescription"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful create batch",
            "schema": {
              "$ref": "#/definitions/BatchStatus"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "422": {
            "description": "Invalid data in request",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/batches/": {
      "get": {
        "tags": [
          "Batch"
        ],
        "summary": "Gets batches",
        "operationId": "getBatches",
        "responses": {
          "200": {
            "description": "Successful get batches",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BatchStatus"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/batches/{batchName}": {
      "get": {
        "tags": [
          "Batch"
        ],
        "summary": "Gets batch",
        "operationId": "getBatch",
        "parameters": [
          {
            "type": "string",
            "description": "Name of batch",
            "name": "batchName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful get batch",
            "schema": {
              "$ref": "#/definitions/BatchStatus"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Batch"
        ],
        "summary": "Delete batch",
        "operationId": "deleteBatch",
        "parameters": [
          {
            "type": "string",
            "description": "Name of batch",
            "name": "batchName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful delete batch",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/batches/{batchName}/jobs/{jobName}": {
      "get": {
        "tags": [
          "Batch"
        ],
        "summary": "Gets batch job",
        "operationId": "getBatchJob",
        "parameters": [
          {
            "type": "string",
            "description": "Name of batch",
            "name": "batchName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Name of job",
            "name": "jobName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful get job",
            "schema": {
              "$ref": "#/definitions/JobStatus"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/batches/{batchName}/jobs/{jobName}/stop": {
      "post": {
        "tags": [
          "Batch"
        ],
        "summary": "Stop batch job",
        "operationId": "stopBatchJob",
        "parameters": [
          {
            "type": "string",
            "description": "Name of batch",
            "name": "batchName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Name of job",
            "name": "jobName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful stop batch job",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
    "/batches/{batchName}/stop": {
      "post": {
        "tags": [
          "Batch"
        ],
        "summary": "Stop batch",
        "operationId": "stopBatch",
        "parameters": [
          {
            "type": "string",
            "description": "Name of batch",
            "name": "batchName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful stop batch",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BatchScheduleDescription": {
      "description": "BatchScheduleDescription holds description about batch scheduling job",
      "type": "object",
      "required": [
        "jobScheduleDescriptions"
      ],
      "properties": {
        "batchId": {
          "description": "Defines a user defined ID of the batch.",
          "type": "string",
          "x-go-name": "BatchId",
          "example": "'batch-id-1'"
        },
        "defaultRadixJobComponentConfig": {
          "$ref": "#/definitions/RadixJobComponentConfig"
        },
        "jobScheduleDescriptions": {
          "description": "JobScheduleDescriptions descriptions of jobs to schedule within the batch",
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobScheduleDescription"
          },
          "x-go-name": "JobScheduleDescriptions"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/common"
    },
    "BatchStatus": {
      "description": "BatchStatus holds general information about batch status",
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "DeploymentName": {
          "description": "DeploymentName for this batch",
          "type": "string"
        },
        "batchId": {
          "description": "Defines a user defined ID of the batch.",
          "type": "string",
          "x-go-name": "BatchId",
          "example": "'batch-id-1'"
        },
        "batchType": {
          "description": "BatchType Single job or multiple jobs batch",
          "type": "string",
          "x-go-name": "BatchType",
          "example": "\"job\""
        },
        "created": {
          "description": "Created timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created"
        },
        "ended": {
          "description": "Ended timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Ended"
        },
        "jobStatuses": {
          "description": "JobStatuses of the jobs in the batch",
          "type": "array",
          "items": {
            "$ref": "#/definitions/JobStatus"
          },
          "x-go-name": "JobStatuses"
        },
        "message": {
          "description": "Message describing the reason for the status",
          "type": "string",
          "x-go-name": "Message"
        },
        "name": {
          "description": "Name of the batch",
          "type": "string",
          "x-go-name": "Name"
        },
        "started": {
          "description": "Started timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Started"
        },
        "status": {
          "description": "Status of the batch.\nWaiting BatchStatusEnumWaiting  Waiting means that all jobs are waiting to start   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nActive BatchStatusEnumActive  Active means that one or more jobs are active   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nCompleted BatchStatusEnumCompleted  Completed means that all jobs have completed (status stopped, succeeded or failed)   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nRunning BatchStatusEnumRunning  Active is a custom status that can only be defined with batchStatusRules in radixconfig\nSucceeded BatchStatusEnumSucceeded  Succeeded is a custom status that can only be defined with batchStatusRules in radixconfig\nFailed BatchStatusEnumFailed  Failed is a custom status that can only be defined with batchStatusRules in radixconfig\nStopping BatchStatusEnumStopping  Stopping is a custom status that can only be defined with batchStatusRules in radixconfig\nStopped BatchStatusEnumStopped  Stopped is a custom status that can only be defined with batchStatusRules in radixconfig",
          "type": "string",
          "enum": [
            "Waiting",
            "Active",
            "Completed",
            "Running",
            "Succeeded",
            "Failed",
            "Stopping",
            "Stopped"
          ],
          "x-go-enum-desc": "Waiting BatchStatusEnumWaiting  Waiting means that all jobs are waiting to start   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nActive BatchStatusEnumActive  Active means that one or more jobs are active   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nCompleted BatchStatusEnumCompleted  Completed means that all jobs have completed (status stopped, succeeded or failed)   This status can also be defined with batchStatusRules in radixconfig, which will override the original meaning\nRunning BatchStatusEnumRunning  Active is a custom status that can only be defined with batchStatusRules in radixconfig\nSucceeded BatchStatusEnumSucceeded  Succeeded is a custom status that can only be defined with batchStatusRules in radixconfig\nFailed BatchStatusEnumFailed  Failed is a custom status that can only be defined with batchStatusRules in radixconfig\nStopping BatchStatusEnumStopping  Stopping is a custom status that can only be defined with batchStatusRules in radixconfig\nStopped BatchStatusEnumStopped  Stopped is a custom status that can only be defined with batchStatusRules in radixconfig",
          "x-go-name": "Status",
          "example": "Waiting"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/v1"
    },
...
    "FailurePolicy": {
      "type": "object",
      "title": "FailurePolicy describes how failed job replicas influence the backoffLimit.",
      "required": [
        "rules"
      ],
      "properties": {
        "rules": {
          "description": "A list of failure policy rules. The rules are evaluated in order.\nOnce a rule matches a job replica failure, the remaining of the rules are ignored.\nWhen no rule matches the failure, the default handling applies - the\ncounter of failures is incremented and it is checked against\nthe backoffLimit.",
          "type": "array",
          "maxItems": 20,
          "items": {
            "$ref": "#/definitions/FailurePolicyRule"
          },
          "x-go-name": "Rules"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/common"
    },
...
    "JobScheduleDescription": {
      "description": "JobScheduleDescription holds description about scheduling job",
      "type": "object",
      "properties": {
        "backoffLimit": {
          "description": "BackoffLimit defines attempts to restart job if it fails. Corresponds to BackoffLimit in K8s.",
          "type": "integer",
          "format": "int32",
          "x-go-name": "BackoffLimit"
        },
        "failurePolicy": {
          "$ref": "#/definitions/FailurePolicy"
        },
        "imageTagName": {
          "description": "ImageTagName defines the image tag name to use for the job image",
          "type": "string",
          "x-go-name": "ImageTagName"
        },
        "jobId": {
          "description": "JobId Optional ID of a job",
          "type": "string",
          "x-go-name": "JobId",
          "example": "'job1'"
        },
        "node": {
          "$ref": "#/definitions/Node"
        },
        "payload": {
          "description": "Payload holding json data to be mapped to component",
          "type": "string",
          "x-go-name": "Payload",
          "example": "{'data':'value'}"
        },
        "resources": {
          "$ref": "#/definitions/Resources"
        },
        "timeLimitSeconds": {
          "description": "TimeLimitSeconds defines maximum job run time. Corresponds to ActiveDeadlineSeconds in K8s.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "TimeLimitSeconds"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/common"
    },
    "JobStatus": {
      "description": "JobStatus holds general information about job status",
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "batchName": {
          "description": "BatchName Optional Batch ID of a job",
          "type": "string",
          "x-go-name": "BatchName",
          "example": "'batch1'"
        },
        "created": {
          "description": "Created timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created"
        },
        "ended": {
          "description": "Ended timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Ended"
        },
        "failed": {
          "description": "The number of times the container for the job has failed.\n+optional",
          "type": "integer",
          "format": "int32",
          "x-go-name": "Failed"
        },
        "jobId": {
          "description": "JobId Optional ID of a job",
          "type": "string",
          "x-go-name": "JobId",
          "example": "'job1'"
        },
        "message": {
          "description": "Message, if any, of the job",
          "type": "string",
          "x-go-name": "Message",
          "example": "\"Error occurred\""
        },
        "name": {
          "description": "Name of the job",
          "type": "string",
          "x-go-name": "Name",
          "example": "calculator"
        },
        "podStatuses": {
          "description": "PodStatuses for each pod of the job",
          "type": "array",
          "items": {
            "$ref": "#/definitions/PodStatus"
          },
          "x-go-name": "PodStatuses"
        },
        "restart": {
          "description": "Timestamp of the job restart, if applied.\n+optional",
          "type": "string",
          "x-go-name": "Restart"
        },
        "started": {
          "description": "Started timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Started"
        },
        "status": {
          "description": "Status of the job\nWaiting JobStatusEnumWaiting  Waiting means that the job is waiting to start\nActive JobStatusEnumActive  Active means that the job is active, but the container is not yet running\nRunning JobStatusEnumRunning  Active means that the job is active and the container is running\nSucceeded JobStatusEnumSucceeded  Succeeded means that the job has completed without errors\nFailed JobStatusEnumFailed  Failed means that the job has failed\nStopping JobStatusEnumStopping  Stopping means that a request to stop the job is sent, but not yet processed by Radix\nStopped JobStatusEnumStopped  Stopped means that the job has been stopped",
          "type": "string",
          "enum": [
            "Waiting",
            "Active",
            "Running",
            "Succeeded",
            "Failed",
            "Stopping",
            "Stopped"
          ],
          "x-go-enum-desc": "Waiting JobStatusEnumWaiting  Waiting means that the job is waiting to start\nActive JobStatusEnumActive  Active means that the job is active, but the container is not yet running\nRunning JobStatusEnumRunning  Active means that the job is active and the container is running\nSucceeded JobStatusEnumSucceeded  Succeeded means that the job has completed without errors\nFailed JobStatusEnumFailed  Failed means that the job has failed\nStopping JobStatusEnumStopping  Stopping means that a request to stop the job is sent, but not yet processed by Radix\nStopped JobStatusEnumStopped  Stopped means that the job has been stopped",
          "x-go-name": "Status",
          "example": "Waiting"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/v1"
    },
    "Node": {
      "description": "Node defines node attributes, where container should be scheduled",
      "type": "object",
      "properties": {
        "gpu": {
          "description": "Defines rules for allowed GPU types.",
          "type": "string",
          "x-go-name": "Gpu"
        },
        "gpuCount": {
          "description": "Defines minimum number of required GPUs.",
          "type": "string",
          "x-go-name": "GpuCount"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/common"
    },
    "PodStatus": {
      "type": "object",
      "title": "PodStatus contains details for the current status of the job's pods.",
      "required": [
        "name"
      ],
      "properties": {
        "containerStarted": {
          "description": "Container started timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "ContainerStarted"
        },
        "created": {
          "description": "Created timestamp",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created"
        },
        "endTime": {
          "description": "The time at which the batch job's pod finishedAt.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "EndTime"
        },
        "exitCode": {
          "description": "Exit status from the last termination of the container",
          "type": "integer",
          "format": "int32",
          "x-go-name": "ExitCode"
        },
        "image": {
          "description": "The image the container is running.",
          "type": "string",
          "x-go-name": "Image",
          "example": "radixdev.azurecr.io/app-server:cdgkg"
        },
        "imageId": {
          "description": "ImageID of the container's image.",
          "type": "string",
          "x-go-name": "ImageId",
          "example": "radixdev.azurecr.io/app-server@sha256:d40cda01916ef63da3607c03785efabc56eb2fc2e0dab0726b1a843e9ded093f"
        },
        "name": {
          "description": "Pod name",
          "type": "string",
          "x-go-name": "Name",
          "example": "server-78fc8857c4-hm76l"
        },
        "podIndex": {
          "description": "The index of the pod in the re-starts",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PodIndex"
        },
        "reason": {
          "description": "A brief CamelCase message indicating details about why the job is in this phase",
          "type": "string",
          "x-go-name": "Reason"
        },
        "replicaStatus": {
          "$ref": "#/definitions/ReplicaStatus"
        },
        "restartCount": {
          "description": "RestartCount count of restarts of a component container inside a pod",
          "type": "integer",
          "format": "int32",
          "x-go-name": "RestartCount"
        },
        "startTime": {
          "description": "The time at which the batch job's pod startedAt",
          "type": "string",
          "format": "date-time",
          "x-go-name": "StartTime"
        },
        "statusMessage": {
          "description": "StatusMessage provides message describing the status of a component container inside a pod",
          "type": "string",
          "x-go-name": "StatusMessage"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/v1"
    },
    "RadixJobComponentConfig": {
      "description": "RadixJobComponentConfig holds description of RadixJobComponent",
      "type": "object",
      "properties": {
        "backoffLimit": {
          "description": "BackoffLimit defines attempts to restart job if it fails. Corresponds to BackoffLimit in K8s.",
          "type": "integer",
          "format": "int32",
          "x-go-name": "BackoffLimit"
        },
        "failurePolicy": {
          "$ref": "#/definitions/FailurePolicy"
        },
        "imageTagName": {
          "description": "ImageTagName defines the image tag name to use for the job image",
          "type": "string",
          "x-go-name": "ImageTagName"
        },
        "node": {
          "$ref": "#/definitions/Node"
        },
        "resources": {
          "$ref": "#/definitions/Resources"
        },
        "timeLimitSeconds": {
          "description": "TimeLimitSeconds defines maximum job run time. Corresponds to ActiveDeadlineSeconds in K8s.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "TimeLimitSeconds"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/models/common"
    },
...
    "Status": {
      "description": "Status is a return value for calls that don't return other objects or when a request returns an error",
      "type": "object",
      "properties": {
        "code": {
          "description": "Suggested HTTP return code for this status, 0 if not set.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 404
        },
        "message": {
          "description": "A human-readable description of the status of this operation.",
          "type": "string",
          "x-go-name": "Message",
          "example": "job job123 is not found"
        },
        "reason": {
          "$ref": "#/definitions/StatusReason"
        },
        "status": {
          "description": "Status of the operation.\nOne of: \"Success\" or \"Failure\".",
          "type": "string",
          "x-go-name": "Status",
          "example": "Failure"
        }
      },
      "x-go-package": "github.com/equinor/radix-job-scheduler/api/errors"
    },
    "StatusReason": {
      "type": "string",
      "x-go-package": "github.com/equinor/radix-job-scheduler/api/errors"
    }
  }
}